pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "aico"
version = "0.5.0"
description = "AICO Monorepo - Unified CLI, Backend, and Shared Libraries"
requires-python = ">=3.13"

# Core dependencies from the 'shared' module
dependencies = [
    "cryptography",
    "keyring",
    "libsql-client>=0.3.1",
    "pyyaml",
    "jsonschema",
    "watchdog",
    "pyzmq>=27.0.1",
    "protobuf>=6.32.0,<7.0.0",
    "platformdirs>=4.0.0",
    "psutil>=6.1.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt>=4.0.1",
    "libsql>=0.1.8",
    "uvicorn>=0.35.0",
    "pynacl>=1.5.0",
]

[project.optional-dependencies]
backend = [
    "chromadb>=1.0.16",
    "duckdb>=1.3.2",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "libsql==0.1.8",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "uvicorn>=0.35.0",
]

cli = [
    "typer>=0.12.3,<1.0.0",
    "rich>=13.7.0,<14.0.0",
    "requests>=2.31.0,<3.0.0",
    "cron-descriptor>=1.4.3",
]

test = [
    "pytest>=8.4.1",
]

[project.scripts]
aico = "cli.aico_main:app"

[tool.setuptools]
packages = ["aico", "cli", "backend", "modelservice"]

[tool.setuptools.package-dir]
aico = "shared/aico"
cli = "cli"
backend = "backend"

[dependency-groups]
test = [
    "pytest-asyncio>=1.1.0",
]
